fix: introduce sdkmantool for java to extend
